requirements.txt
numpy==1.19.4
sklearn==0.0
scikit-learn==0.24.2
torchvision==0.10.0
scipy==1.5.4
matplotlib==3.3.4
pandas==1.1.5
tqdm==4.62.2
nltk==3.6.7
seaborn==1.16.0
openpyxl==3.1.2
tab-transformer-pytorch==0.3.0

